feat/#180 : 패킹 리스트 초대 - api 변경사항 반영 - 유저 권한 검사 로직을 컴포넌트 > 페이지로 분리
