+++ service/packingList/together/index.ts
@@ -410,7 +410,7 @@ export interface GetInvitedOutput {
   success: boolean;
   message: string;
   data: {
-    _id: string; // 함께 패킹리스트 id
+    id: string; // 함께 패킹리스트 id
     title: string; // 함께 패킹리스트 제목
   };
 }
